# Make PostBodyContentTypeParser 1.5x - 10x faster