--- a/lib/rack/contrib/post_body_content_type_parser.rb
+++ b/lib/rack/contrib/post_body_content_type_parser.rb
@@ -1,70 +1,46 @@
 begin
   require 'json'
-rescue LoadError
+rescue LoadError => e
   require 'json/pure'
 end
 
 module Rack
-  # A Rack middleware for making JSON-encoded request bodies available in the
-  # request.params hash. By default it parses POST, PATCH, and PUT requests,
-  # but you can configure it to parse any request type via the :verbs option
+
+  # A Rack middleware for parsing POST/PUT body data when Content-Type is
+  # not one of the standard supported types, like <tt>application/json</tt>.
   #
-  # Examples:
-  #     use Rack::PostBodyContentTypeParser, verbs: %w[POST GET]
+  # TODO: Find a better name.
+  #
   class PostBodyContentTypeParser
+
+    # Constants
+    #
     CONTENT_TYPE = 'CONTENT_TYPE'.freeze
-    DEFAULT_VERBS = %w[POST PATCH PUT].freeze
-    JSON_CONTENT_TYPE = /json/.freeze
-    DEFAULT_JSON_PARSER = ->(body) { JSON.parse(body, create_additions: false) }
+    POST_BODY = 'rack.input'.freeze
+    FORM_INPUT = 'rack.request.form_input'.freeze
+    FORM_HASH = 'rack.request.form_hash'.freeze
 
-    module Matchers
-      # Backport Ruby 2.4's regexp matcher, so Ruby >= 2.4 runs at top speed
-      unless ''.respond_to?(:match?)
-        refine String do
-          def match?(regex)
-            self =~ regex
-          end
-        end
-      end
+    # Supported Content-Types
+    #
+    APPLICATION_JSON = 'application/json'.freeze
 
-      # env[CONTENT_TYPE] can be nil, so nil must handle #match? in this scope
-      refine NilClass do
-        def match?(_)
-          false
-        end
-      end
-    end
-
-    using Matchers
-
-    def initialize(app, config = {}, &json_parser)
+    def initialize(app, &block)
       @app = app
-      @verbs = config[:verbs] || DEFAULT_VERBS
-      @json_parser = json_parser || DEFAULT_JSON_PARSER
+      @block = block || Proc.new { |body| JSON.parse(body, :create_additions => false) }
     end
 
     def call(env)
-      if @verbs.include?(env[Rack::REQUEST_METHOD]) &&
-         env[CONTENT_TYPE].match?(JSON_CONTENT_TYPE)
-
-        write_json_body_to(env)
+      if Rack::Request.new(env).media_type == APPLICATION_JSON && (body = env[POST_BODY].read).length != 0
+        env[POST_BODY].rewind # somebody might try to read this stream
+        env.update(FORM_HASH => @block.call(body), FORM_INPUT => env[POST_BODY])
       end
       @app.call(env)
     rescue JSON::ParserError
-      Rack::Response.new('failed to parse body as JSON', 400).finish
+      bad_request('failed to parse body as JSON')
     end
 
-    private
-
-    def write_json_body_to(env)
-      body = env[Rack::RACK_INPUT]
-      return unless (body_content = body.read) && !body_content.empty?
-
-      body.rewind # somebody might try to read this stream
-      env.update(
-        Rack::RACK_REQUEST_FORM_HASH => @json_parser.call(body_content),
-        Rack::RACK_REQUEST_FORM_INPUT => body
-      )
+    def bad_request(body = 'Bad Request')
+      [ 400, { 'Content-Type' => 'text/plain', 'Content-Length' => body.bytesize.to_s }, [body] ]
     end
   end
 end