This is a non-breaking (I hope!) refactor of PostBodyContentTypeParser, aimed at improving performance. It uses a couple of tactics to make things faster:

1. Skip parsing GET, OPTIONS and other bodiless HTTP requests by default. Because the HTTP spec disallows these request types from having a body, this change shouldn't affect anyone's app — although it is the change I'm most worried about. Seems worth the risk, though, because it improves performance of GET requests by 10x.

2. For POST, PATCH, and PUT requests, (a) read the content-type from `env` directly, instead of instantiating a `Rack::Request`, and (b) use Ruby 2.4's `String#match?` operator when available. These changes speed up the requests that do get parsed by 1.5x.

This PR also makes the middleware more liberal about what content-types it will accept. Instead of just `application/json`, it matches any content-type that includes `json`. I made this change because I want to support the JSON:API content-type: https://jsonapi.org/format/#content-negotiation-clients

If this PR gets merged, I figured it'd be on me to fix any bugs, so I prioritized making the code readable (to me) over minimizing the number of lines I changed. I'm happy to try again with an eye toward minimizing the diff, though, if that would be preferable to you @mpalmer.

Here's the benchmark I was running to measure performance changes:

```ruby
require 'bundler/setup'
require 'rack'
require 'rack/contrib'
require 'benchmark/ips'

App = Rack::Builder.new {
  use Rack::PostBodyContentTypeParser
  run lambda { |env| [200, { 'Content-Length' => '2' }, ['ok']] }
}.to_app

body = StringIO.new('{"key":"value"}')
config = { input: body, 'CONTENT_TYPE' => 'application/json' }
get_env = Rack::MockRequest.env_for('/', config)
post_env = Rack::MockRequest.env_for('/', config.merge(method: 'POST'))

Benchmark.ips do |bm|
  bm.report 'GET' do
    App.call(get_env)
  end

  bm.report 'POST' do
    App.call(post_env)
  end

  bm.compare!
end
```

I ran it first from the master branch and then again from my fork. Here are the results:

```
MASTER
Warming up --------------------------------------
        GET (master)    19.126k i/100ms
       POST (master)    19.025k i/100ms
Calculating -------------------------------------
        GET (master)    204.358k (± 4.3%) i/s -      1.033M in   5.064006s
       POST (master)    196.148k (± 4.0%) i/s -    989.300k in   5.051819s

Comparison:
        GET (master):   204358.1 i/s
       POST (master):   196147.9 i/s - same-ish: difference falls within error

PATCHED
Warming up --------------------------------------
       GET (patched)   135.596k i/100ms
      POST (patched)    26.123k i/100ms
Calculating -------------------------------------
       GET (patched)      2.137M (± 0.8%) i/s -     10.712M in   5.012443s
      POST (patched)    290.596k (± 1.7%) i/s -      1.463M in   5.035550s

Comparison:
       GET (patched):  2137227.5 i/s
      POST (patched):   290596.3 i/s - 7.35x  slower
```

As you can see, GET performance jumps from 200k IPS to 2M IPS, and POST performance jumps from 200K IPS to 290K IPS.