@mpalmer 2019-05-29
Thanks for submitting this.  Apologies for the delay in responding.

I'm always a big fan of performance improvements, and your code (and tests) look really good overall.  With a bit more backwards compatibility, I think we're in business.

As you say, the HTTP spec says "no bodies in GET/HEAD requests", however I'm willing to bet that *someone* out there who uses `PostBodyContentTypeParser` is using it to parse GET request bodies, because the universe tends towards maximum chaos.  Changing the set of content types matched, especially with a regex match, feels like it might catch some people unawares, too.

Frankly, given that you've given the middleware a (much needed and appreciated) thorough overhaul, I'm inclined to think it'll work out better to make a new middleware, maybe just called `JSONBodyParser`, that behaves in the way you've implemented, and deprecate `PostBodyContentTypeParser`, with a view to removing it when we next need a major release.

I think it makes the most sense to make the content types it'll work on configurable, so you can use it on your JSON:API types if you like, via regex, while keeping the default more conservative.