@chrisfrank 2019-06-04
Thanks for being open to this, @mpalmer. I agree with your compatibility concerns, and I think adding a new middleware is a solid way of working around them. I'll aim to finish a revision to this week that works as you've suggested.