@chrisfrank 2019-07-28
Sorry for the delay here, @mpalmer. This should be ready to go now. As per our discussion, this revision deprecates PostBodyContentTypeParser and adds JSONBodyParser.

JSONBodyParser accepts a `verbs` keyword to control which HTTP verbs match, a `media` keyword to control which media types match, and (like PostBodyContentTypeParser) a block to control how the JSON body gets parsed.

I made the benchmark more realistic by removing the body from the GET requests. This change makes PostBodyContentParser perform better, although JSONBodyParser is still 7x faster.

For POST requests, it's faster to use a regex media type matcher instead of the default string matcher, because the regex matcher doesn't need to call `Rack::MediaType.type`. We could consider making the default matcher be the regex `/application\/json/` instead of the string 'application/json', but I defer to your judgment about whether that's too loose as a default.

Here are the current benchmark results:
```
Warming up --------------------------------------
          Old [GET]     35.636k i/100ms
          Old [POST]    17.856k i/100ms
          New [GET]    166.323k i/100ms
 New [POST] [string]    19.988k i/100ms
  New [POST] [regex]    31.639k i/100ms
Calculating -------------------------------------
          Old [GET]     402.355k (± 1.0%) i/s -      2.031M in   5.048907s
          Old [POST]    196.174k (± 1.9%) i/s -    982.080k in   5.008029s
          New [GET]       2.652M (± 1.1%) i/s -     13.306M in   5.018897s
 New [POST] [string]    219.365k (± 1.9%) i/s -      1.099M in   5.013326s
  New [POST] [regex]    350.833k (± 1.8%) i/s -      1.772M in   5.051889s

Comparison:
          New [GET] :  2651515.0 i/s
          Old [GET] :   402355.2 i/s - 6.59x  slower
  New [POST] [regex]:   350833.1 i/s - 7.56x  slower
 New [POST] [string]:   219364.7 i/s - 12.09x  slower
          Old [POST]:   196173.6 i/s - 13.52x  slower
```

And the benchmark code...
```ruby
require 'bundler/setup'
require 'benchmark/ips'
require 'rack'
require 'rack/contrib'

app = ->(env) { [200, { 'Content-Length': '2'}, ['OK']] }

Old = Rack::Builder.new {
  use Rack::PostBodyContentTypeParser
  run app
}.to_app

New = Rack::Builder.new {
  use Rack::JSONBodyParser
  run app
}.to_app

NewWithRegexp = Rack::Builder.new {
  use Rack::JSONBodyParser, media: /application\/json/
  run app
}.to_app

body = StringIO.new('{"key":"value"}')
config = { 'CONTENT_TYPE' => 'application/json; charset=utf-8' }
get_env = Rack::MockRequest.env_for('/', config)
post_env = Rack::MockRequest.env_for('/', config.merge(method: 'POST', input: body))

Benchmark.ips do |bm|
  bm.report('Old [GET] ') { Old.call(get_env) }
  bm.report('Old [POST]') { Old.call(post_env) }
  bm.report('New [GET] ') { New.call(get_env) }
  bm.report('New [POST] [string]') { New.call(post_env) }
  bm.report('New [POST] [regex]') { NewWithRegexp.call(post_env) }

  bm.compare!
end
```