@ioquatix 2020-03-31
Apparently Elastic Search uses a body with GET requests :/ Just FYI.